feat: Enhance account management with validation and soft delete functionality
- Added Zod validation for profile updates in AccountController.
- Refactored delete method in AccountRepository to implement soft delete by updating user status instead of permanent deletion.
- Improved error handling in AccountService for user updates and deletions, ensuring proper authorization checks.
- Expanded user model to include isDeleted and deletedAt fields, with a pre-find middleware to exclude deleted users from queries.
- Updated routes to include new endpoints for updating and deleting accounts.

+++ src/feature/account/service/account.service.ts
@@ -1,17 +1,37 @@
 import { User } from "../../../types/models/user";
-import { BadRequestError } from "../../../utils/error.utils";
+import { BadRequestError, UnauthorizedError } from "../../../utils/error.utils";
 import accountRepository from "../repository/account.repository";
 import { FilterQuery, ObjectId } from "mongoose";
 
 class AccountService {
-  async getUserById(id: ObjectId | string) {
-    return await accountRepository.findById(id);
+  /**
+   * Get user by ID
+   * @param id - User ID
+   * @param select - Select fields
+   * @returns User
+   */
+  async getUserById(
+    id: ObjectId | string,
+    select: string = "-password -refreshToken"
+  ) {
+    return await accountRepository.findById(id, select);
   }
 
+  /**
+   * Get user by email
+   * @param email - User email
+   * @returns User
+   */
   async getUserByEmail(email: string) {
     return await accountRepository.findByEmail(email);
   }
 
+  /**
+   * Get users by filter
+   * @param filter - Filter query
+   * @param select - Select fields
+   * @returns Users
+   */
   async getUsersByFilter(
     filter: FilterQuery<User>,
     select?: string | Record<string, number>
@@ -19,10 +39,21 @@ class AccountService {
     return await accountRepository.findByFilter(filter, select);
   }
 
+  /**
+   * Create user
+   * @param userData - User data to create
+   * @returns User
+   */
   async createUser(userData: Partial<User>) {
     return await accountRepository.create(userData);
   }
 
+  /**
+   * Update user
+   * @param id - User ID
+   * @param updateData - User data to update
+   * @returns User
+   */
   async updateUser(
     id: ObjectId | string,
     updateData: Partial<User>
@@ -31,22 +62,64 @@ class AccountService {
       throw new BadRequestError("User ID is required");
     }
 
+    const user = await this.getUserById(id, "_id role");
+
+    if (!user) {
+      throw new BadRequestError("User not found");
+    }
+
+    if (updateData.refreshToken) {
+      throw new UnauthorizedError(
+        "You are not authorized to update refresh token"
+      );
+    }
+
+    if (updateData.role && user.role !== "admin") {
+      throw new UnauthorizedError("You are not authorized to update role");
+    }
+
     return await accountRepository.update(id, updateData);
   }
 
-  async deleteUser(id: string) {
+  /**
+   * Delete user
+   * @param id - User ID
+   * @returns User ID
+   */
+  async deleteUser(id: ObjectId | string) {
+    // check if user is admin
+    const user = await this.getUserById(id, "_id role");
+
+    if (!user) {
+      throw new BadRequestError("User not found");
+    }
+
     return await accountRepository.delete(id);
   }
 
+  /**
+   * Get all users
+   * @returns All users
+   */
   async getAllUsers() {
     return await accountRepository.findAll();
   }
 
+  /**
+   * Check if user exists
+   * @param email - User email
+   * @returns boolean
+   */
   async checkUserExists(email: string): Promise<boolean> {
     const user = await this.getUserByEmail(email);
     return !!user;
   }
 
+  /**
+   * Get user profile
+   * @param id - User ID
+   * @returns User profile
+   */
   async getUserProfile(id: string) {
     const user = await this.getUserById(id);
     if (!user) {
@@ -56,6 +129,11 @@ class AccountService {
     return user;
   }
 
+  /**
+   * Logout user and remove refresh token
+   * @param userId - User ID
+   * @returns User ID
+   */
   async logout(userId: ObjectId | string): Promise<ObjectId | string | null> {
     console.log(userId);
